+++ nautilus_trader/common/providers.py
@@ -152,10 +152,14 @@ class InstrumentProvider:
                 await self.load_all_async(self._filters)
             elif self._load_ids_on_start:
                 instrument_ids = [
-                    InstrumentId.from_str(i)
+                    i if isinstance(i, InstrumentId) else InstrumentId.from_str(i)
                     for i in self._load_ids_on_start
-                    if not isinstance(i, InstrumentId)
                 ]
+
+                instruments_str = "".join([i.value for i in instrument_ids])
+                filters_str = "..." if not self._filters else f" with filters {self._filters}..."
+                self._log.info(f"Loading instruments {instruments_str}{filters_str}")
+
                 await self.load_ids_async(instrument_ids, self._filters)
             self._log.info(f"Loaded {self.count} instruments")
         else: