+++ RELEASES.md
@@ -12,6 +12,7 @@ None
 None
 
 ### Fixes
+- Fixed loading specific instrument IDs for `InstrumentProviderConfig`
 - Fixed reconcile open orders and account websocket message for dYdX (#2039), thanks @davidsblom
 - Fixed market order `avg_px` for Polymarket trade reports
 - Fixed Betfair clients keepalive (#2040), thanks @limx0